Refactor packages
Improve package structure
Remove unnecessary packages
