+++ src/main/java/de/tum/www1/orion/OrionStartupProjectRefreshActivity.kt
@@ -6,16 +6,16 @@ import com.intellij.openapi.startup.StartupActivity
 import com.intellij.testFramework.runInEdtAndGet
 import de.tum.www1.orion.connector.client.JavaScriptConnector
 import de.tum.www1.orion.connector.ide.vcs.submit.ChangeSubmissionContext
+import de.tum.www1.orion.exercise.registry.BrokenRegistryLinkException
+import de.tum.www1.orion.exercise.registry.OrionGlobalExerciseRegistryService
+import de.tum.www1.orion.exercise.registry.OrionProjectRegistryStateService
+import de.tum.www1.orion.exercise.registry.OrionStudentExerciseRegistry
 import de.tum.www1.orion.messaging.OrionIntellijStateNotifier
+import de.tum.www1.orion.settings.OrionSettingsProvider
 import de.tum.www1.orion.ui.util.BrokenLinkWarning
-import de.tum.www1.orion.util.OrionSettingsProvider
 import de.tum.www1.orion.util.appService
-import de.tum.www1.orion.util.registry.BrokenRegistryLinkException
-import de.tum.www1.orion.util.registry.OrionGlobalExerciseRegistryService
-import de.tum.www1.orion.util.registry.OrionProjectRegistryStateService
-import de.tum.www1.orion.util.registry.OrionStudentExerciseRegistry
 import de.tum.www1.orion.util.service
-import de.tum.www1.orion.vcs.OrionGitUtil
+import de.tum.www1.orion.vcs.OrionGitAdapter
 
 class OrionStartupProjectRefreshActivity : StartupActivity {
 
@@ -52,7 +52,7 @@ class OrionStartupProjectRefreshActivity : StartupActivity {
     private fun prepareExercise(registry: OrionStudentExerciseRegistry, project: Project) {
         registry.exerciseInfo?.let { exerciseInfo ->
             project.messageBus.syncPublisher(OrionIntellijStateNotifier.INTELLIJ_STATE_TOPIC).openedExercise(exerciseInfo.exerciseId, exerciseInfo.view)
-            OrionGitUtil.updateExercise(project)
+            OrionGitAdapter.updateExercise(project)
         }
     }
 }
