+++ src/main/java/de/tum/www1/orion/exercise/registry/OrionExerciseRegistry.kt
@@ -1,4 +1,4 @@
-package de.tum.www1.orion.util.registry
+package de.tum.www1.orion.exercise.registry
 
 import de.tum.www1.orion.dto.ProgrammingExercise
 import de.tum.www1.orion.enumeration.ExerciseView
