+++ src/main/java/de/tum/www1/orion/exercise/registry/OrionStudentExerciseRegistry.kt
@@ -1,4 +1,4 @@
-package de.tum.www1.orion.util.registry
+package de.tum.www1.orion.exercise.registry
 
 import com.intellij.openapi.components.ServiceManager
 import com.intellij.openapi.project.Project
