+++ src/main/java/de/tum/www1/orion/ui/browser/BrowserWebView.java
@@ -6,9 +6,9 @@ import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.project.Project;
 import com.intellij.util.messages.Topic;
+import de.tum.www1.orion.settings.OrionSettingsProvider;
 import de.tum.www1.orion.ui.OrionRouter;
 import de.tum.www1.orion.ui.util.UrlAccessForbiddenWarning;
-import de.tum.www1.orion.util.OrionSettingsProvider;
 import javafx.application.Platform;
 import javafx.concurrent.Worker;
 import javafx.embed.swing.JFXPanel;
