+++ src/main/resources/META-INF/plugin.xml
@@ -42,13 +42,13 @@
     
     <extensions defaultExtensionNs="com.intellij">
         <toolWindow id="Artemis" anchor="right" icon="/toolWindowIcon.png" factoryClass="de.tum.www1.orion.ui.browser.BrowserFactory" />
-        <applicationService serviceInterface="de.tum.www1.orion.vcs.CredentialsService"
-                            serviceImplementation="de.tum.www1.orion.vcs.CredentialsInjectorService"/>
+        <applicationService serviceInterface="de.tum.www1.orion.vcs.OrionGitCredentialsService"
+                            serviceImplementation="de.tum.www1.orion.vcs.OrionGitCredentialsInjectorService"/>
         <projectService serviceInterface="de.tum.www1.orion.ui.OrionRouter"
                         serviceImplementation="de.tum.www1.orion.ui.OrionRouterService"/>
         <projectService serviceImplementation="de.tum.www1.orion.ui.browser.Browser" />
-        <projectService serviceInterface="de.tum.www1.orion.util.registry.OrionStudentExerciseRegistry"
-                        serviceImplementation="de.tum.www1.orion.util.registry.DefaultOrionStudentExerciseRegistry"/>
+        <projectService serviceInterface="de.tum.www1.orion.exercise.registry.OrionStudentExerciseRegistry"
+                        serviceImplementation="de.tum.www1.orion.exercise.registry.DefaultOrionStudentExerciseRegistry"/>
         <projectService serviceInterface="de.tum.www1.orion.connector.client.JavaScriptConnector"
                         serviceImplementation="de.tum.www1.orion.connector.client.ArtemisClientConnector"/>
         <projectService serviceInterface="de.tum.www1.orion.connector.ide.shared.IOrionSharedUtilConnector"
@@ -61,8 +61,8 @@
                         serviceImplementation="de.tum.www1.orion.connector.ide.exercise.OrionExerciseConnector"/>
         <projectService serviceImplementation="de.tum.www1.orion.connector.ide.vcs.submit.ChangeSubmissionContext"/>
         <postStartupActivity implementation="de.tum.www1.orion.OrionStartupProjectRefreshActivity"/>
-        <applicationService serviceInterface="de.tum.www1.orion.util.OrionSettingsProvider"
-                            serviceImplementation="de.tum.www1.orion.util.OrionSettingsProviderService"/>
+        <applicationService serviceInterface="de.tum.www1.orion.settings.OrionSettingsProvider"
+                            serviceImplementation="de.tum.www1.orion.settings.OrionSettingsProviderService"/>
         <projectConfigurable instance="de.tum.www1.orion.ui.settings.OrionPluginSettings" id="de.tum.www1.orion.ui.settings"
                              nonDefaultProject="true" groupId="tools" bundle="i18n.OrionBundle"
                              key="orion.settings.displayName" />
@@ -70,13 +70,13 @@
         <programRunner implementation="de.tum.www1.orion.build.OrionProgramRunner" />
         <projectService serviceInterface="de.tum.www1.orion.build.OrionTestParser"
                         serviceImplementation="de.tum.www1.orion.build.OrionTestInterceptor"/>
-        <projectService serviceInterface="de.tum.www1.orion.util.registry.OrionInstructorExerciseRegistry"
-                        serviceImplementation="de.tum.www1.orion.util.registry.DefaultOrionInstructorExerciseRegistry"/>
+        <projectService serviceInterface="de.tum.www1.orion.exercise.registry.OrionInstructorExerciseRegistry"
+                        serviceImplementation="de.tum.www1.orion.exercise.registry.DefaultOrionInstructorExerciseRegistry"/>
         <projectService serviceImplementation="de.tum.www1.orion.build.instructor.OrionInstructorBuildUtil"/>
         <applicationService serviceInterface="de.tum.www1.orion.util.storage.LocalBrowserStorageProvider"
                             serviceImplementation="de.tum.www1.orion.util.storage.LocalBrowserStorage"/>
-        <projectService serviceImplementation="de.tum.www1.orion.util.registry.OrionProjectRegistryStateService"/>
-        <applicationService serviceImplementation="de.tum.www1.orion.util.registry.OrionGlobalExerciseRegistryService"/>
+        <projectService serviceImplementation="de.tum.www1.orion.exercise.registry.OrionProjectRegistryStateService"/>
+        <applicationService serviceImplementation="de.tum.www1.orion.exercise.registry.OrionGlobalExerciseRegistryService"/>
     </extensions>
 
     <actions>
